updated the certificate model.

+++ src/main/java/edu/team5/wright_time/model/repository/CertificationRepository.java
@@ -3,6 +3,6 @@ package edu.team5.wright_time.model.repository;
 import edu.team5.wright_time.model.entity.Certification;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CertificationRepository extends CrudRepository<Certification,Integer> {
+public interface CertificationRepository extends CrudRepository <Certification, Long> {
 
 }
